fix: deadlocks that happens on task switches
We might lock the `MMListInner` in some operation when we do a
task switch. Then in the idle task, we try to switch page table.
But that's impossible because we shouldn't do anything that could
possibly put us in a sleeping state.

Also, when we return from `schedule()` in `CondVar::wait()`, we might
be waken up from some signals. We should clean ourselves from the
waiters queue. Otherwise, we might be accidentally waken up by some
other tasks and ends up in some inconsistent task scheduler state.