feat: set ap cpus online in scheduler!
There are still some bugs waiting to be fixed. The biggest of all is
that multiple processes call `exit` at the same time will cause
deadlock. This is because we hold `current.inner`, then `init.inner` and
then `child.inner`. And if our children are doing `exit` simultaneously,
they might hold their `current.inner` first and spin on `init.inner`.

This causes the deadlock.