runtime: add trace logs and fix few bugs
Add tracing logs in Runtime::enter and other critical points.

Pass trace_scheduler feature down to eonix_runtime crate, fixing the
problem that the feature is not working.

When the task is blocked, we set CURRENT_TASK to None as well.

In early initialization stage, the stack is placed in identically mapped
physical address. VirtIO driver might try converting the given buffer
paths back to physical ones, which will generate errors. So BSP and AP
should allocate an another stack and switch to it. We use TaskContext
for the fix.

Signed-off-by: greatbridf <greatbridf@icloud.com>