riscv64, trap: remove load_interrupt_stack impl
This is used only by Thread when we enter user execution context, when
we need to save the "interrupt stack" to the local CPU so we can get the
information needed to capture the trap.

We need to support nested captured trap returns. So instead of setting
that manually, we save the needed information when trap_return() is
called (since we have precisely the trap context needed) and restore it
after the trap is captured.

Signed-off-by: greatbridf <greatbridf@icloud.com>