riscv64, trap: rework to fix nested captured traps
The previous implementation has some bugs inside that will cause kernel
space nested traps to lose some required information:

- In kernel mode, trap contexts are saved above the current stack frame
  without exception, which is not what we want. We expect to read the
  trap data in the CAPTURED context.
- The capturer task context is not saved as well, which will mess up the
  nested traps completely.
- We are reading page fault virtual addresses in TrapContext::trap_type,
  which won't work since if the inner trap is captured, and the outer
  trap interleaves with the trap_type() call, we will lose the stval
  data in the inner trap.

The solution is to separate our "normal" trap handling procedure out of
captured trap handling procedure. We swap the stvec CSR when we set up
captured traps and restore it afterwards so the two approach don't have
to tell then apart in trap entries. Then, we can store the TrapContext
pointer in sscratch without having to distinguish between trap handling
types. In the way, we keep the procedure simple.

The register stval is saved together with other registers to be used in
page faults.

Signed-off-by: greatbridf <greatbridf@icloud.com>