task, thread: working version of threads
We've got everything done in order to make the system run.

Add Thread::contexted to load the context needed for the thread to run.
Wrap the Thread::real_run() with contexted(stackful(...)) in
Thread::run().

We would use this for now. Later, we will make the thread completely
asynchronous. This way we don't have to change its interface then.

Signed-off-by: greatbridf <greatbridf@icloud.com>