task: fix stackful waker implementation
The current implementation use the WokenUp object to detect whether the
stackful task is woken up somewhere. This is WRONG since we might lose
wakeups as the runtime have no idea what we have done. If someone wakes
us up, the task won't be enqueued so we will never have a second chance
to get to the foreground.

The fix is to use Arc<Task> to create a waker and check whether the task
is ready each time we get back to the stackful poll loop.

Signed-off-by: greatbridf <greatbridf@icloud.com>